+++ CHANGES.md
@@ -59,6 +59,10 @@
 
   - <https://github.com/georust/gdal/pull/265>
 
+- Remove `PartialEq` from `GdalError`
+
+  - <https://github.com/georust/gdal/pull/286>
+
 - Prevent SIGGEGV when reading a string array on an MD Array that is not of type string.
 
   - <https://github.com/georust/gdal/pull/284>